Fix a bug: kthread_task_init() corrupted the stack.
When the kernel starts, its starts with its esp pointing to the top of
the available kern_stack. A little later in the initialization the
cosecd task is prepared. Part of the preparation is setting up an
"interrupted state" on the new task stack, which is the same
`kern_stack`. By a sheer luck, at that point Ubuntu builds use just
enough stack to give the initialization enough space. macOS builds using
i686-elf-gcc from Homebrew were not so lucky and corrupted the stack,
leading to a fatal page fault during boot.

TODO: figure out safe stack arrangement or just use separate stacks
during boot and for cosecd and switch them.